Updating how profile object is passed
