+++ messaging_controller.py
@@ -1,7 +1,6 @@
 from flask import Flask, url_for
 app = Flask(__name__)
 from flask import render_template
-from users import *
 from profile import *
