+++ profile_model.py
@@ -1,0 +1,14 @@
+from flask import Flask, url_for
+app = Flask(__name__)
+from flask import render_template
+from canvas import *
+
+
+
+class Profile():
+  pass
+    
+  
+
+
+
