+++ templates/post.html
@@ -1,10 +1,10 @@
 <!DOCTYPE html>
   
 
-<article id="post_{{posts[i].id}}" class="post_box">
+<article id="post_{{profile.posts[i].id}}" class="post_box">
   <div class="profile_name">
     <div class="profile-pic">
-      {% if(posts[i].title == 'Admin' or posts[i].title == 'Announcement' ) %}
+      {% if(profile.posts[i].title == 'Admin' or profile.posts[i].title == 'Announcement' ) %}
         <figure class="thumbnail"><img alt="placeholder"class="img-fluid rounded-circle" src="../static/images/p4h.png"/></figure>
       {% else %}
         <figure class="thumbnail"> <img alt="placeholder"class="img-fluid rounded-circle" src="../static/images/me.png"/> </figure>
@@ -12,47 +12,47 @@
     </div>
     <div class="col-10">
       <header class="text-left">
-        <figcaption class="comment-user"><b>{{posts[i].title}}</b></figcaption>
+        <figcaption class="comment-user"><b>{{profile.posts[i].title}}</b></figcaption>
         <time class="comment-date" datetime="{{date}}"><i class="fa fa-clock-o"></i> {{date}}</time>
       </header>
     </div>
   </div>
         <div class="post">
           <div class="">
-            {{posts[i].message}}
+            {{profile.posts[i].message}}
           </div>
         <hr>
-          {% if( (comments[posts[i].id | string] | length) > 0 ) %}
+          {% if( (profile.profile.comments[profile.posts[i].id | string] | length) > 0 ) %}
           <div class="text-center">
-              <p class=""><a href="" name="{{posts[i].id}}" id="view_more-{{posts[i].id}}" class="profile_button view_more">View Comments</a></p>
+              <p class=""><a href="" name="{{profile.posts[i].id}}" id="view_more-{{profile.posts[i].id}}" class="profile_button view_more">View Comments</a></p>
           </div>
           {%endif%}              
-          <div id="comments-{{posts[i].id}}" style="display:none;">
-          <label class = "comment_label" for="from"><b>Comments</b></label>
-          {% for j in range(comments[posts[i].id | string] | length) | reverse %}
+          <div id="profile.comments-{{profile.posts[i].id}}" style="display:none;">
+          <label class = "comment_label" for="from"><b>profile.comments</b></label>
+          {% for j in range(profile.comments[profile.posts[i].id | string] | length) | reverse %}
           <div class="post_comment">
             <figure class=" col-2"><img alt="placeholder"class="img-fluid rounded-circle" src="../static/images/me.png"/></figure>
             <div class="word_bubble col-10">
-              <p class=""><b>{{comments[posts[i].id | string][j].user["display_name"]}}:</b><br>
-              {{comments[posts[i].id | string][j].message}}</p>
+              <p class=""><b>{{profile.comments[profile.posts[i].id | string][j].user["display_name"]}}:</b><br>
+              {{profile.comments[profile.posts[i].id | string][j].message}}</p>
             </div>
           </div>
           {% endfor %}
           </div>
-          <div id="reply_div-{{posts[i].id}}" class="reply_div "> 
+          <div id="reply_div-{{profile.posts[i].id}}" class="reply_div "> 
             <div class="col-8">
-              <textarea class= "text_box" name="message" id="textbox_reply-{{posts[i].id}}" style="" onkeyup="Expand(this);" size="5" placeholder="Comment"></textarea><span class="upload_icon oi oi-cloud-camera" aria-hidden="true"></span>
+              <textarea class= "text_box" name="message" id="textbox_reply-{{profile.posts[i].id}}" style="" onkeyup="Expand(this);" size="5" placeholder="Comment"></textarea><span class="upload_icon oi oi-cloud-camera" aria-hidden="true"></span>
             </div>
-            <a href="" name="{{posts[i].id}}" id="reply-{{posts[i].id}}"  class="col-4 reply_button btn-sm"><i class="fa fa-reply"></i> Reply</a>
+            <a href="" name="{{profile.posts[i].id}}" id="reply-{{profile.posts[i].id}}"  class="col-4 reply_button btn-sm"><i class="fa fa-reply"></i> Reply</a>
           </div>
         </div>
       </article>
 <script>
-// handle posts
+// handle profile.posts
 $(document).on("click", ".view_more", function(e) { 
   e.preventDefault();
   var id = e.currentTarget.name;
   console.log("view more: ", id);
-  $("#comments-"+id).css("display",""); 
+  $("#profile.comments-"+id).css("display",""); 
 });
 </script>
